+++ packages/core/src/impl-store.ts
@@ -72,12 +72,10 @@ async function fetchOrDeployGeneric<T extends Deployment, U extends T = T>(
       );
       if (updated !== stored) {
         if (merge && deployment.merge) {
-          // only check primary addresses for clashes, since the address could already exist in an allAddresses field
-          // but the above updated and stored objects are different instances representing the same entry
-          await checkForAddressClash(provider, data, updated, false);
+          await checkForAddressClash(provider, data, updated, true);
           deployment.merge(updated);
         } else {
-          await checkForAddressClash(provider, data, updated, true);
+          await checkForAddressClash(provider, data, updated, false);
           deployment.set(updated);
         }
         await manifest.write(data);
@@ -172,12 +170,17 @@ export async function fetchOrDeployGetDeployment<T extends ImplDeployment>(
 const implLens = (versionWithoutMetadata: string) =>
   lens(`implementation ${versionWithoutMetadata}`, 'implementation', data => ({
     get: () => data.impls[versionWithoutMetadata],
-    set: (value?: ImplDeployment) => (data.impls[versionWithoutMetadata] = value),
-    merge: (value?: ImplDeployment) => {
+    set: (value?: ImplDeployment & RemoteDeploymentId) => (data.impls[versionWithoutMetadata] = value),
+    merge: (value?: ImplDeployment & RemoteDeploymentId) => {
       const existing = data.impls[versionWithoutMetadata];
       if (existing !== undefined && value !== undefined) {
         const { address, allAddresses } = mergeAddresses(existing, value);
-        data.impls[versionWithoutMetadata] = { ...existing, address, allAddresses };
+        data.impls[versionWithoutMetadata] = {
+          ...existing,
+          address,
+          allAddresses,
+          remoteDeploymentId: value.remoteDeploymentId,
+        };
       } else {
         data.impls[versionWithoutMetadata] = value;
       }
@@ -223,14 +226,23 @@ async function checkForAddressClash(
   provider: EthereumProvider,
   data: ManifestData,
   updated: Deployment & RemoteDeploymentId,
-  checkAllAddresses: boolean,
+  merge: boolean,
 ): Promise<void> {
-  const clash = lookupDeployment(data, updated.address, checkAllAddresses);
-  if (clash !== undefined) {
-    if (await isDevelopmentNetwork(provider)) {
+  let clash;
+  const findClash = () => lookupDeployment(data, updated.address, !merge);
+
+  if (await isDevelopmentNetwork(provider)) {
+    // Look for clashes so that we can delete deployments from older runs.
+    // `merge` only checks primary addresses for clashes, since the address could already exist in an allAddresses field
+    // but the updated and stored objects are different instances representing the same entry.
+    clash = findClash();
+    if (clash !== undefined) {
       debug('deleting a previous deployment at address', updated.address);
       clash.set(undefined);
-    } else {
+    }
+  } else if (!merge) {
+    clash = findClash();
+    if (clash !== undefined) {
       const existing = clash.get();
       // it's a clash if there is no deployment id or if deployment ids don't match
       if (
@@ -238,13 +250,12 @@ async function checkForAddressClash(
         (existing.remoteDeploymentId === undefined || existing.remoteDeploymentId !== updated.remoteDeploymentId)
       ) {
         throw new Error(
-          `The following deployment clashes with an existing one at ${updated.address}\n\n` +
-            JSON.stringify(updated, null, 2) +
-            `\n\n`,
+          `The deployment clashes with an existing one at ${updated.address}.\n\n` +
+            `Deployments: ${JSON.stringify({ existing, updated }, null, 2)}\n\n`,
         );
       }
     }
-  }
+  } // else, merge indicates that the user is force-importing or redeploying an implementation, so we simply allow merging the entries
 }
 
 function lookupDeployment(